% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{bls_get_cache_dir}
\alias{bls_get_cache_dir}
\title{Get the current BLSloadR Cache Directory}
\usage{
bls_get_cache_dir()
}
\value{
A character string of the cache path.
}
\description{
Displays the path currently being used for caching. This will check the
`BLS_CACHE_DIR` environment variable, falling back to the default
system cache directory if the variable is not set.
}
