% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{check_bls_cache_env}
\alias{check_bls_cache_env}
\title{Check if Global Caching is Enabled via Environment Variable}
\usage{
check_bls_cache_env()
}
\value{
Logical value indicating if the environment variable USE_BLS_CACHE is one of TRUE, YES, or 1
}
\description{
Check if Global Caching is Enabled via Environment Variable
}
\keyword{internal}
