% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_overview.R
\name{display_in_popup}
\alias{display_in_popup}
\title{Display text content in popup window.}
\usage{
display_in_popup(content, series_id)
}
\arguments{
\item{content}{Character. Text content to display}

\item{series_id}{Two-letter series ID for a BLS time series to render in the display.}
}
\value{
No object returned, called to render content in popup.
}
\description{
Helper function used to display content from 'bls_overview' in a popup window.
}
