% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{download_bls_files}
\alias{download_bls_files}
\title{Helper function for downloading and tracking BLS files}
\usage{
download_bls_files(
  urls,
  suppress_warnings = TRUE,
  user_agent = NULL,
  cache = check_bls_cache_env()
)
}
\arguments{
\item{urls}{Named or unnamed character vector of URLs to download. If unnamed, names will be auto-generated from basenames.}

\item{suppress_warnings}{Logical. If TRUE, suppress individual download warnings}

\item{user_agent}{An optional character string to pass to the USER_AGENT HTML header}

\item{cache}{Logical. If TRUE, uses local persistent caching.}
}
\value{
Named list of bls_data objects
}
\description{
This function is used to pass multiple URLs at the Bureau of Labor Statistics into 'fread_bls()'
}
