% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bls_object.R
\name{get_bls_data}
\alias{get_bls_data}
\title{Extract data from BLS data object}
\usage{
get_bls_data(bls_obj)
}
\arguments{
\item{bls_obj}{A bls_data_collection object or raw data}
}
\value{
The data component
}
\description{
Extract data from BLS data object
}
