% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_bls_diagnostics}
\alias{get_bls_diagnostics}
\title{Get download diagnostics from BLS data object}
\usage{
get_bls_diagnostics(bls_obj)
}
\arguments{
\item{bls_obj}{A bls_data_collection object}
}
\value{
List of download diagnostics from a bls_data_collection object.
}
\description{
This is a helper function to extract the download diagnostics element of a 'bls_data_collection' object.
}
