% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_bls_summary}
\alias{get_bls_summary}
\title{Get summary information from BLS data object}
\usage{
get_bls_summary(bls_obj)
}
\arguments{
\item{bls_obj}{A bls_data_collection object}
}
\value{
List of summary information
}
\description{
This is a helper function to extract the summary element of a 'bls_data_collection' object. This containes the number of files downloaded, the number of files with potential warnings, and the total number of warnings.
}
