% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bls_object.R
\name{has_bls_issues}
\alias{has_bls_issues}
\title{Check if BLS data object has issues}
\usage{
has_bls_issues(bls_obj)
}
\arguments{
\item{bls_obj}{A BLS data object}
}
\value{
Logical indicating if there were any issues
}
\description{
Check if BLS data object has issues
}
