% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces_helpers.R
\name{list_ces_states}
\alias{list_ces_states}
\title{List Available States for CES Data}
\usage{
list_ces_states()
}
\value{
A character vector of available state/territory abbreviations
}
\description{
Lists all available U.S. states and territories that can be used with the 
`states` parameter in `get_ces()` function.
}
\examples{
# See all available states
list_ces_states()

# Use with get_ces
# ces_data <- get_ces(states = c("MA", "NY"))  # All industries for these states
}
