% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bls_object.R
\name{print.bls_data_collection}
\alias{print.bls_data_collection}
\title{Print method for bls_data_collection objects}
\usage{
\method{print}{bls_data_collection}(x, ...)
}
\arguments{
\item{x}{A bls_data_collection object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for bls_data_collection objects
}
