% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_BLS.R
\name{read_bls_excel}
\alias{read_bls_excel}
\title{Download BLS Excel Data}
\usage{
read_bls_excel(url, user_agent = NULL, verbose = FALSE, ...)
}
\arguments{
\item{url}{Character string. URL to the BLS .xlsx or .xls file.}

\item{user_agent}{An optional character string to pass to the USER_AGENT HTML header.}

\item{verbose}{Logical. If TRUE, prints diagnostic messages.}

\item{...}{Additional arguments passed to readxl::read_excel (e.g., sheet, range).}
}
\value{
A data.frame or NULL if the download or read fails.
}
\description{
Download BLS Excel Data
}
\examples{
\dontrun{
# Download BLS Alternative MEasures History
salt_url <- "https://www.bls.gov/lau/stalt-moave.xlsx"
salt_data <- read_bls_excel(salt_url, skip = 1)

}

}
