% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_ces_helpers.R
\name{show_national_ces_options}
\alias{show_national_ces_options}
\title{Show National CES Dataset Options and Usage Examples}
\usage{
show_national_ces_options()
}
\value{
Prints formatted information to the console.
}
\description{
This function provides a comprehensive overview of the national CES dataset
filtering options available in get_national_ces(), including examples
of how to use each option.
}
\examples{
show_national_ces_options()

}
