% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{create_bls_object}
\alias{create_bls_object}
\title{Create a BLS data object with diagnostics}
\usage{
create_bls_object(
  data,
  downloads,
  data_type = "BLS",
  processing_steps = character(0)
)
}
\arguments{
\item{data}{The processed data (data.table/data.frame)}

\item{downloads}{List of download results from fread_bls()}

\item{data_type}{Character string describing the type of BLS data (e.g., "CES", "JOLTS", "CPS")}

\item{processing_steps}{Character vector describing processing steps applied}
}
\value{
A bls_data_collection object
}
\description{
This is a helper function to create a list with the additional class 'bls_data_collection' containing data downloaded form the U.S. Bureau of Labor Statistics as well as diagnostic details about the download. It is used invisibly in the package to bundle information about file downloads.
}
