% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bls_object.R
\name{download_bls_files}
\alias{download_bls_files}
\title{Helper function for downloading and tracking BLS files}
\usage{
download_bls_files(urls, suppress_warnings = FALSE)
}
\arguments{
\item{urls}{Named character vector of URLs to download}

\item{suppress_warnings}{Logical. If TRUE, suppress individual download warnings}
}
\value{
Named list of bls_data objects
}
\description{
Helper function for downloading and tracking BLS files
}
