% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{download_bls_files}
\alias{download_bls_files}
\title{Helper function for downloading and tracking BLS files}
\usage{
download_bls_files(urls, suppress_warnings = TRUE)
}
\arguments{
\item{urls}{Named character vector of URLs to download}

\item{suppress_warnings}{Logical. If TRUE, suppress individual download warnings}
}
\value{
Named list of bls_data objects
}
\description{
This function is used to pass multiple URLs at the Bureau of Labor Statistics into 'fread_bls()'
}
