% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_BLS.R
\name{fread_bls}
\alias{fread_bls}
\title{Download BLS Time Series Data}
\usage{
fread_bls(url)
}
\arguments{
\item{url}{Character string. URL to the BLS flat file}
}
\value{
A data.table containing the downloaded data
}
\description{
This function uses data.table::fread and httr to retrieve data from BLS
time series flat files, adding appropriate headers to avoid 403 errors.
}
\examples{
\dontrun{
data <- fread_bls("https://download.bls.gov/pub/time.series/sm/sm.series")
}
}
