% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_bls.R
\name{fread_bls}
\alias{fread_bls}
\title{Download BLS Time Series Data}
\usage{
fread_bls(url, verbose = FALSE, cache = check_bls_cache_env())
}
\arguments{
\item{url}{Character string. URL to the BLS flat file}

\item{verbose}{Logical. If TRUE, prints additional messages during file read and processing.}

\item{cache}{Logical. If TRUE, uses local persistent caching.}
}
\value{
A named list with the data and diagnostics.
}
\description{
This function downloads a tab-delimited BLS flat file, incorporating 
diagnostic information about the file and returning an object with the
bls_data class that can be used in the BLSloadR package.
}
