% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_helpers.R
\name{get_bls_data}
\alias{get_bls_data}
\title{Extract data from BLS data object}
\usage{
get_bls_data(bls_obj)
}
\arguments{
\item{bls_obj}{A bls_data_collection object or raw data}
}
\value{
The data component of a 'bls_data_collection' object as a data frame.
}
\description{
This is a helper function to extract the data element of a 'bls_data_collection' object.
}
