% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bls_object.R
\name{get_bls_summary}
\alias{get_bls_summary}
\title{Get summary information from BLS data object}
\usage{
get_bls_summary(bls_obj)
}
\arguments{
\item{bls_obj}{A bls_data_collection object}
}
\value{
List of summary information
}
\description{
Get summary information from BLS data object
}
